# optimize  batch_and_prepare_render_phase