--- a/crates/bevy_pbr/src/render/mesh.rs
+++ b/crates/bevy_pbr/src/render/mesh.rs
@@ -472,13 +472,11 @@ impl GetBatchData for MeshPipeline {
     fn get_batch_data(
         (mesh_instances, lightmaps): &SystemParamItem<Self::Param>,
         entity: &QueryItem<Self::Data>,
-    ) -> (Self::BufferData, Option<Self::CompareData>) {
-        let mesh_instance = mesh_instances
-            .get(entity)
-            .expect("Failed to find render mesh instance");
+    ) -> Option<(Self::BufferData, Option<Self::CompareData>)> {
+        let mesh_instance = mesh_instances.get(entity)?;
         let maybe_lightmap = lightmaps.render_lightmaps.get(entity);
 
-        (
+        Some((
             MeshUniform::new(
                 &mesh_instance.transforms,
                 maybe_lightmap.map(|lightmap| lightmap.uv_rect),
@@ -488,7 +486,7 @@ impl GetBatchData for MeshPipeline {
                 mesh_instance.mesh_asset_id,
                 maybe_lightmap.map(|lightmap| lightmap.image),
             )),
-        )
+        ))
     }
 }