--- a/crates/bevy_render/src/batching/mod.rs
+++ b/crates/bevy_render/src/batching/mod.rs
@@ -2,7 +2,6 @@ use bevy_ecs::{
     component::Component,
     entity::Entity,
     prelude::Res,
-    query::{QueryFilter, QueryItem, ReadOnlyQueryData},
     system::{Query, ResMut, StaticSystemParam, SystemParam, SystemParamItem},
 };
 use bevy_utils::nonmax::NonMaxU32;
@@ -58,8 +57,6 @@ impl<T: PartialEq> BatchMeta<T> {
 /// items.
 pub trait GetBatchData {
     type Param: SystemParam + 'static;
-    type Data: ReadOnlyQueryData;
-    type Filter: QueryFilter;
     /// Data used for comparison between phase items. If the pipeline id, draw
     /// function id, per-instance data buffer dynamic offset and this data
     /// matches, the draws can be batched.
@@ -73,16 +70,13 @@ pub trait GetBatchData {
     /// for the `CompareData`.
     fn get_batch_data(
         param: &SystemParamItem<Self::Param>,
-        query_item: &QueryItem<Self::Data>,
+        query_item: &Entity,
     ) -> Option<(Self::BufferData, Option<Self::CompareData>)>;
 }
 
 /// Batch the items in a render phase. This means comparing metadata needed to draw each phase item
 /// and trying to combine the draws into a batch.
-pub fn batch_and_prepare_render_phase<
-    I: CachedRenderPipelinePhaseItem,
-    F: GetBatchData<Data = Entity>,
->(
+pub fn batch_and_prepare_render_phase<I: CachedRenderPipelinePhaseItem, F: GetBatchData>(
     gpu_array_buffer: ResMut<GpuArrayBuffer<F::BufferData>>,
     mut views: Query<&mut RenderPhase<I>>,
     param: StaticSystemParam<F::Param>,
@@ -90,7 +84,7 @@ pub fn batch_and_prepare_render_phase<
     let gpu_array_buffer = gpu_array_buffer.into_inner();
     let system_param_item = param.into_inner();
 
-    let process_item = |item: &mut I| {
+    let mut process_item = |item: &mut I| {
         let (buffer_data, compare_data) = F::get_batch_data(&system_param_item, &item.entity())?;
         let buffer_index = gpu_array_buffer.push(buffer_data);
 
@@ -105,49 +99,9 @@ pub fn batch_and_prepare_render_phase<
         }
     };
 
-    apply_batch::<I, F>(&mut views, process_item);
-}
-
-/// Batch the items in a render phase with Query to fetch its required data. If Query is unnecessary ,using [`batch_and_prepare_render_phase`] will be more efficient, Not used yet
-pub fn batch_and_prepare_render_phase_with_query<
-    I: CachedRenderPipelinePhaseItem,
-    F: GetBatchData,
->(
-    gpu_array_buffer: ResMut<GpuArrayBuffer<F::BufferData>>,
-    mut views: Query<&mut RenderPhase<I>>,
-    param: StaticSystemParam<F::Param>,
-    query: Query<F::Data, F::Filter>,
-) {
-    let gpu_array_buffer = gpu_array_buffer.into_inner();
-    let system_param_item = param.into_inner();
-
-    let process_item = |item: &mut I| {
-        let batch_query_item = query.get(item.entity()).ok()?;
-
-        let (buffer_data, compare_data) = F::get_batch_data(&system_param_item, &batch_query_item)?;
-        let buffer_index = gpu_array_buffer.push(buffer_data);
-
-        let index = buffer_index.index.get();
-        *item.batch_range_mut() = index..index + 1;
-        *item.dynamic_offset_mut() = buffer_index.dynamic_offset;
-
-        if I::AUTOMATIC_BATCHING {
-            compare_data.map(|compare_data| BatchMeta::new(item, compare_data))
-        } else {
-            None
-        }
-    };
-
-    apply_batch::<I, F>(&mut views, process_item);
-}
-
-fn apply_batch<I: CachedRenderPipelinePhaseItem, F: GetBatchData>(
-    views: &mut Query<&mut RenderPhase<I>>,
-    mut process: impl FnMut(&mut I) -> Option<BatchMeta<<F as GetBatchData>::CompareData>>,
-) {
-    for mut phase in views {
+    for mut phase in &mut views {
         let items = phase.items.iter_mut().map(|item| {
-            let batch_data = process(item);
+            let batch_data = process_item(item);
             (item.batch_range_mut(), batch_data)
         });
         items.reduce(|(start_range, prev_batch_meta), (range, batch_meta)| {
@@ -160,6 +114,7 @@ fn apply_batch<I: CachedRenderPipelinePhaseItem, F: GetBatchData>(
         });
     }
 }
+
 pub fn write_batched_instance_buffer<F: GetBatchData>(
     render_device: Res<RenderDevice>,
     render_queue: Res<RenderQueue>,