@re0312 2024-01-18
> Could you update the "solution" part of the description with your latest changes?

done!